add post and comment validation to show 404 page if not exist

+++ handlers/deletecomment.py
@@ -7,6 +7,10 @@ class DeleteComment(BlogHandler):
     def get(self, comment_id):
         comment = Comment.get_by_id(comment_id)
 
+        if not comment:
+            self.error(404)
+            return
+
         if self.user:
             if comment.user.key().id() == self.user.key().id():
                 comment.delete()
