+++ handlers/deletepost.py
@@ -9,6 +9,10 @@ class DeletePost(BlogHandler):
     def get(self, post_id):
         post = Post.get_by_id(post_id)
 
+        if not post:
+            self.error(404)
+            return
+
         if self.user:
             if post.is_owner(self.user):
                 post.delete()
